Refactor: move methods to ctrl_file.cpp
