+++ variables.hpp
@@ -885,7 +885,6 @@ void addefmap(int = 0, int = 0, int = 0, int = 0, int = 0, int = 0);
 void anime_halt();
 void animeblood(int = 0, int = 0, int = 0);
 void animeload(int = 0, int = 0);
-void arrayfile(bool, const std::string&, const fs::path&);
 void auto_turn(int = 0);
 void atxinit();
 void begintempinv();
@@ -915,7 +914,6 @@ void dmgplayer(int = 0, int = 0);
 void eatstatus(curse_state_t, int);
 void efllistadd(int = 0, int = 0, int = 0, int = 0, int = 0, int = 0);
 void exittempinv();
-void fileadd(const fs::path& filepath, int = 0);
 void fix_input_chat(std::string&);
 void fix_input_chat2(std::string&);
 void fixaiact(int = 0);
